Stop pretending like readers are Not Our Problem
If we want to be able to support multiple transport stacks in a sensible manner, we're going to need to make it our problem. This takes a bit of refactoring, because Rust is a good and sensible language.

I still cannot believe that Rc<Impl> is an actual pattern in Rust. I love C++.